updated the dashboard
