+++ src/main/resources/templates/chores/edit.html
@@ -19,18 +19,18 @@
     <label for="childAssigned">Assigned to:</label>
     <input type="text" id="childAssigned" name="childAssigned" th:value="${chore.childAssigned}" th:field="${chore.childAssigned}" class="form-control" required />
   </div>
-  <div class="form-group">
-    <label for="dueDay">Due Day:</label>
-    <input type="number" id="dueDay" name="dueDay" th:value="${chore.dueDay}" th:field="${chore.dueDay}" class="form-control" required />
-  </div>
 
   <div class="form-group">
-    <label for="dueDate">Due Date:</label>
+    <label for="dueDate">Date:</label>
     <input type="number" id="dueDate" name="dueDate" th:value="${chore.dueDate}" th:field="${chore.dueDate}" class="form-control" required />
   </div>
   <div class="form-group">
     <label for="rewardPoints">Rewards Value:</label>
     <input type="number" id="rewardPoints" name="rewardPoints" th:value="${chore.rewardPoints}" th:field="${chore.rewardPoints}" class="form-control" required />
+  </div>
+  <div class="form-group">
+    <label for="dueDay">Due Day:</label>
+    <input type="number" id="dueDay" name="dueDay" th:value="${chore.dueDay}" th:field="${chore.dueDay}" class="form-control" required />
   </div>
   <button type="submit" class="btn btn-primary">Save Changes</button>
   <a th:href="@{/rewards}" class="btn btn-secondary">Cancel</a>
