+++ src/main/resources/templates/dashboard.html
@@ -110,22 +110,6 @@
 </script>
 
 
-<!-- Avatar selection script -->
-<!--script>
-    function selectAvatar(avatarUrl) {
-        // Update the avatar on the page
-        document.getElementById("selectedAvatar").querySelector("img").src = "/images/" + {avatar.png, avatar2.png};
-
-        // Send an AJAX request to update the avatar in the backend
-        fetch("/updateAvatar", {
-            method: "POST",
-            headers: {
-                "Content-Type": "application/json"
-            },
-            body: JSON.stringify({ avatarUrl: avatarUrl, avatarId: avatarId })
-        });
-    }
-</-script-->
 
 <!-- Chores and Rewards -->
