+++ src/main/resources/templates/dashfragment.html
@@ -41,7 +41,10 @@
     </div--->
 
     <!-- Points Display -->
-    <h2>Points: <span th:text="${child.points}"></span></h2>
+    <div class="points">
+      <h2>Points: <span th:text="${child.points}"></span></h2>
+    </div>
+
 
     <!-- Redeem Reward Button -->
     <div class="reward-button">
